requirements.txt
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
tenacity==8.5.0
pytest==8.2.0